backend/requirements.txt
# Framework principal
Django==5.1.3
djangorestframework==3.15.2

# Base de données MySQL
mysqlclient==2.2.6

# Authentification JWT
djangorestframework_simplejwt==5.3.1
PyJWT==2.10.1

# CORS pour frontend/backend
django-cors-headers==4.6.0

# Interface admin améliorée
django-jazzmin==3.0.1

# Configuration environnement
python-dotenv==1.0.1
python-decouple==3.8

# Manipulation d'images
pillow==11.0.0

# Outils utiles
pytz==2024.2
sqlparse==0.5.2

# Dépendances de base
asgiref==3.8.1
tzdata==2024.2 